refractor(): functions into class TSP
